+++ androidgraphextension/src/main/kotlin/tech/schoolhelper/androidgraphextension/RectExtension.kt
@@ -29,6 +29,19 @@ fun Rect.contains(event: MotionEvent) = this.contains(event.x, event.y)
 	* @return true iff (x,y) are contained by the rectangle, where containment
 	*              means left <= x < right and top <= y < bottom
 	*/
+fun Rect.contains(x: Double, y: Double) = this.contains(x.toInt(), y.toInt())
+
+/**
+	* Returns true if (x,y) is inside the rectangle. The left and top are
+	* considered to be inside, while the right and bottom are not. This means
+	* that for a x,y to be contained: left <= x < right and top <= y < bottom.
+	* An empty rectangle never contains any point.
+	*
+	* @param x The X coordinate of the point being tested for containment
+	* @param y The Y coordinate of the point being tested for containment
+	* @return true iff (x,y) are contained by the rectangle, where containment
+	*              means left <= x < right and top <= y < bottom
+	*/
 fun Rect.contains(x: Float, y: Float) = this.contains(x.toInt(), y.toInt())
 
 /**